fix accesses

+++ application/classes/controller/authentication.php
@@ -14,18 +14,21 @@ class Controller_Authentication extends Controller {
             $user instanceof Model_User;
             $date = date_parse($user->registration_date);
             $date_time = mktime($date['hour'], $date['minute'], $date['second'], $date['month'], $date['day'], $date['year']);
-            if(time() - $date_time > 24*60*60) {
+            if(false && (time() - $date_time > 24*60*60)) {
                 $user->delete();
                 Plussia_Info::factory('OutOfAuthTime');
             } else {
                 $mp = Model_MailProve::findOneBy(array('password'=>$_GET['password']));
-                $user->addRole(Model_Role::get('login'));
-                $user->addRole(Model_Role::get('registration'));
-                $user->addRole(Model_Role::get('regstep1'));
                 $user->set('afterlogin', 'registration');
                 $user->save();
+                Plussia_Dispatcher::setUser($user);
                 $mp && $mp->delete();
-                Plussia_Info::factory('AuthenticationSuccess', array('name' => $user->getUserData()->name));
+
+                $afterlogin = ($user->afterlogin && $user->afterlogin != '') ? $user->afterlogin : $user->username;
+                Plussia_Dispatcher::updateActive(true);
+                header("Location: /" . $afterlogin);
+                die;
+                //Plussia_Info::factory('AuthenticationSuccess', array('name' => $user->getUserData()->name));
             }
         }
         Plussia_Error::call(2);
